requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
gurobipy>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Utilities
psutil>=5.9.0

# UI (for later)
gradio>=3.50.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0